server/requirements.txt
faiss_cpu==1.7.4
fastapi==0.108.0
numpy==1.23.5
pdfminer.six==20231228
pydantic==2.5.3
transformers==4.36.2
python-multipart
gensim
scikit-learn